+++ app/db/models/user.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Integer, String, BigInteger
 from sqlalchemy.orm import relationship
 
 from app.db.models.base import Base
@@ -10,8 +10,6 @@ __all__ = ['User']
 class User(Base):
     __tablename__ = 'users'
 
-    tg_id = Column(Integer, primary_key=True)
+    tg_id = Column(BigInteger, primary_key=True)
     name = Column(String(50), nullable=False)
     timezone = Column(Integer, nullable=False)
-
-
